# docs/swagger.yaml
basePath: /
definitions:
...
  model.CreatePOClientRequest:
    properties:
      delivery:
        type: string
      file:
        type: string
      id_mitra:
        type: integer
      items:
        items:
          $ref: '#/definitions/model.CreatePOClientItemRequest'
        minItems: 1
        type: array
      payment_term:
        type: string
      penanggung_jawab:
        items:
          $ref: '#/definitions/model.CreatePenanggungJawabRequest'
        minItems: 1
        type: array
      po_number:
        type: string
      season:
        type: string
      tanggal:
        type: string
    required:
    - delivery
    - id_mitra
    - items
    - penanggung_jawab
    - po_number
    - tanggal
    type: object
...
  model.CreatePOInternalRequest:
    properties:
      cpo:
        type: string
      currency:
        type: string
      id_pr_internal:
        type: integer
      items:
        items:
          $ref: '#/definitions/model.CreatePOInternalItemRequest'
        minItems: 1
        type: array
      nama_po:
        type: string
      ship_date:
        type: string
      supplier_addr:
        type: string
      supplier_contact:
        type: string
      supplier_email:
        type: string
      supplier_fax:
        type: string
      supplier_name:
        type: string
      supplier_telp:
        type: string
      tanggal:
        type: string
      term:
        type: string
    required:
    - currency
    - id_pr_internal
    - items
    - nama_po
    - ship_date
    - supplier_name
    - tanggal
    type: object
...
  model.CreatePRInternalRequest:
    properties:
      departemen:
        type: string
      id_wo:
        type: integer
      items:
        items:
          $ref: '#/definitions/model.CreatePRInternalItemRequest'
        minItems: 1
        type: array
      nama:
        type: string
      projek:
        type: string
      tanggal:
        type: string
      vendor_address:
        type: string
      vendor_name:
        type: string
      vendor_telp:
        type: string
    required:
    - departemen
    - id_wo
    - items
    - nama
    - projek
    - tanggal
    - vendor_name
    type: object
...
  model.POClientDetailSuccessDoc:
    properties:
      data:
        $ref: '#/definitions/model.POClientDetailResponse'
      message:
        example: po client retrieved
        type: string
      status:
        example: success
        type: string
    type: object
...
  model.POClientListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/model.POClientListItem'
        type: array
      pagination:
        $ref: '#/definitions/model.PaginationMeta'
    type: object
  model.POClientListSuccessDoc:
    properties:
      data:
        $ref: '#/definitions/model.POClientListResponse'
      message:
        example: po clients retrieved
        type: string
      status:
        example: success
        type: string
    type: object
...
  model.POClientSuccessDoc:
    properties:
      data:
        $ref: '#/definitions/model.POClientResponse'
      message:
        example: po client created
        type: string
      status:
        example: success
        type: string
    type: object
  model.POInternalDetailSuccessDoc:
    properties:
      data:
        $ref: '#/definitions/model.POInternalResponse'
      message:
        example: po internal retrieved
        type: string
      status:
        example: success
        type: string
    type: object
...
  model.POInternalListSuccessDoc:
    properties:
      data:
        $ref: '#/definitions/model.POInternalListResponse'
      message:
        example: po internals retrieved
        type: string
      status:
        example: success
        type: string
    type: object
...
  model.POInternalSuccessDoc:
    properties:
      data:
        $ref: '#/definitions/model.POInternalResponse'
      message:
        example: po internal created
        type: string
      status:
        example: success
        type: string
    type: object
  model.PRInternalDetailSuccessDoc:
    properties:
      data:
        $ref: '#/definitions/model.PRInternalResponse'
      message:
        example: pr internal retrieved
        type: string
      status:
        example: success
        type: string
    type: object
...
  model.PRInternalListSuccessDoc:
    properties:
      data:
        $ref: '#/definitions/model.PRInternalListResponse'
      message:
        example: pr internals retrieved
        type: string
      status:
        example: success
        type: string
    type: object
...
  model.PRInternalStatusSuccessDoc:
    properties:
      data:
        $ref: '#/definitions/model.PRInternalStatusResponse'
      message:
        example: pr internal approved
        type: string
      status:
        example: success
        type: string
    type: object
  model.PRInternalSuccessDoc:
    properties:
      data:
        $ref: '#/definitions/model.PRInternalResponse'
      message:
        example: pr internal created
        type: string
      status:
        example: success
        type: string
    type: object
...
  model.TransactionErrorDoc:
    properties:
      error:
        $ref: '#/definitions/model.TransactionErrorDetail'
      message:
        example: related data not found
        type: string
      status:
        example: error
        type: string
    type: object
  model.TransactionValidationErrorDoc:
    properties:
      error:
        items:
          $ref: '#/definitions/response.ValidationErrorItem'
        type: array
      message:
        example: bad request
        type: string
      status:
        example: error
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@permatatex.local
    name: API Support
  description: API documentation for Permatatex IT Inventory System backend.
  license:
    name: MIT
  termsOfService: https://swagger.io/terms/
  title: Permatatex IT Inventory API
  version: "1.0"
paths:
...
  /api/v1/po-clients:
    get:
      description: Returns a paginated list of PO client headers for transaction screens.
      parameters:
      - description: Page (default 1)
        in: query
        name: page
        type: integer
      - description: Limit (default 20)
        in: query
        name: limit
        type: integer
      - description: Search by PO number, season, or mitra name
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.POClientListSuccessDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.TransactionErrorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.TransactionErrorDoc'
      security:
      - BearerAuth: []
      summary: List PO Clients
      tags:
      - Transaction Documents
    post:
      consumes:
      - application/json
      description: Create a PO client document with nested items and penanggung jawab
        in a single transaction.
      parameters:
      - description: PO client payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.CreatePOClientRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.POClientSuccessDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.TransactionValidationErrorDoc'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/model.TransactionErrorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.TransactionErrorDoc'
      security:
      - BearerAuth: []
      summary: Create PO Client
      tags:
      - Transaction Documents
  /api/v1/po-clients/{id}:
    get:
      description: Returns a single PO client with nested items and penanggung jawab.
      parameters:
      - description: PO Client ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.POClientDetailSuccessDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.TransactionErrorDoc'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.TransactionErrorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.TransactionErrorDoc'
      security:
      - BearerAuth: []
      summary: Get PO Client Detail
      tags:
      - Transaction Documents
    put:
      consumes:
      - application/json
      description: Replace a PO client header, items, and penanggung jawab in a single
        transaction.
      parameters:
      - description: PO Client ID
        in: path
        name: id
        required: true
        type: integer
      - description: PO client payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.CreatePOClientRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.POClientSuccessDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.TransactionValidationErrorDoc'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.TransactionErrorDoc'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/model.TransactionErrorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.TransactionErrorDoc'
      security:
      - BearerAuth: []
      summary: Update PO Client
      tags:
      - Transaction Documents
  /api/v1/po-internals:
    get:
      description: Returns a paginated list of PO internal headers.
      parameters:
      - description: Page (default 1)
        in: query
        name: page
        type: integer
      - description: Limit (default 20)
        in: query
        name: limit
        type: integer
      - description: Search by PO name, supplier, or CPO
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.POInternalListSuccessDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.TransactionErrorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.TransactionErrorDoc'
      security:
      - BearerAuth: []
      summary: List PO Internals
      tags:
      - Transaction Documents
    post:
      consumes:
      - application/json
      description: Create a PO internal document with nested items in a single transaction.
      parameters:
      - description: PO internal payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.CreatePOInternalRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.POInternalSuccessDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.TransactionValidationErrorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.TransactionErrorDoc'
      security:
      - BearerAuth: []
      summary: Create PO Internal
      tags:
      - Transaction Documents
  /api/v1/po-internals/{id}:
    get:
      description: Returns a single PO internal with nested items.
      parameters:
      - description: PO Internal ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.POInternalDetailSuccessDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.TransactionErrorDoc'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.TransactionErrorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.TransactionErrorDoc'
      security:
      - BearerAuth: []
      summary: Get PO Internal Detail
      tags:
      - Transaction Documents
  /api/v1/pr-internals:
    get:
      description: Returns a paginated list of PR internal headers.
      parameters:
      - description: Page (default 1)
        in: query
        name: page
        type: integer
      - description: Limit (default 20)
        in: query
        name: limit
        type: integer
      - description: Search by name, vendor, or projek
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.PRInternalListSuccessDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.TransactionErrorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.TransactionErrorDoc'
      security:
      - BearerAuth: []
      summary: List PR Internals
      tags:
      - Transaction Documents
    post:
      consumes:
      - application/json
      description: Create a PR internal document with nested items in a single transaction.
      parameters:
      - description: PR internal payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/model.CreatePRInternalRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.PRInternalSuccessDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.TransactionValidationErrorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.TransactionErrorDoc'
      security:
      - BearerAuth: []
      summary: Create PR Internal
      tags:
      - Transaction Documents
  /api/v1/pr-internals/{id}:
    get:
      description: Returns a single PR internal with nested items.
      parameters:
      - description: PR Internal ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.PRInternalDetailSuccessDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.TransactionErrorDoc'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.TransactionErrorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.TransactionErrorDoc'
      security:
      - BearerAuth: []
      summary: Get PR Internal Detail
      tags:
      - Transaction Documents
  /api/v1/pr-internals/{id}/approve:
    patch:
      description: Manager approval endpoint that only changes PR internal status
        and audit fields.
      parameters:
      - description: PR Internal ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.PRInternalStatusSuccessDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.TransactionErrorDoc'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.TransactionErrorDoc'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/model.TransactionErrorDoc'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.TransactionErrorDoc'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/model.TransactionErrorDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.TransactionErrorDoc'
      security:
      - BearerAuth: []
      summary: Approve PR Internal
      tags:
      - Transaction Documents
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"